server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "motor>=3.7.1",
    "pymupdf-layout>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
